+++ src/main/java/com/everyonewaiter/adapter/config/persistence/JpaConfiguration.java
@@ -1,4 +1,4 @@
-package com.everyonewaiter.adapter.persistence;
+package com.everyonewaiter.adapter.config.persistence;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.persistence.EntityManager;
