+++ src/main/java/com/everyonewaiter/adapter/config/security/SecurityConfiguration.java
@@ -1,4 +1,4 @@
-package com.everyonewaiter.adapter.security;
+package com.everyonewaiter.adapter.config.security;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
